Refactor Pixel Distorsion Shader Components
- Optimize shader performance with improved texture handling
- Add null checks and window existence checks in scene component
- Implement proper texture disposal in pixel distorsion component
- Simplify shader fragment and vertex code
- Update import paths for shader-related files
- Remove unnecessary commented code and debug statements
- Enhance type safety and memoization in shader components
